add fontawesome svg, moving file from profile ui to oma
