+++ assets/js/components/App/App.js
@@ -1,5 +1,4 @@
 import React, {Component} from 'react';
-import {hot} from 'react-hot-loader';
 import {connect} from 'react-redux';
 
 class App extends Component {
@@ -14,10 +13,4 @@ class App extends Component {
     }
 }
 
-App = connect()(App);
-
-if (!process.env.NODE_ENV || process.env.NODE_ENV === 'development') {
-    App = hot(module)(App);
-}
-
-export default App;
+export default connect()(App);
