+++ assets/js/components/OnboardingSteps/Welcome.js
@@ -3,7 +3,7 @@ import {injectIntl} from 'react-intl';
 
 const Welcome = ({intl}) => {
     return (
-        <div>
+        <div className="onboarding-welcome">
             <h2>{intl.formatMessage({id: 'onboarding.welcome.heading'})}</h2>
             <p>{intl.formatMessage({id: 'onboarding.welcome.text'})}</p>
         </div>
