+++ assets/js/components/OnboardingSteps/index.js
@@ -1,0 +1,13 @@
+import StepButtons from './StepButtons';
+import Welcome from './Welcome';
+import PersonalInformation from './PersonalInformation';
+import CreatePassword from './CreatePassword';
+import Settings from './Settings';
+
+export {
+    StepButtons,
+    Welcome,
+    PersonalInformation,
+    CreatePassword,
+    Settings,
+}
