+++ assets/js/index.js
@@ -12,7 +12,7 @@ import App from './components/App/App';
 import configureStore from './root/store';
 
 import MainPage from './components/Main';
-import UserOnboarding from './onboarding';
+import UserOnboarding from './components/Onboarding';
 
 const intlLocaleData = getIntlLocaleData();
 addLocaleData(intlLocaleData);
