+++ omahelsinki/static/css/_omahelsinki.scss
@@ -1,3 +1,8 @@
+/* custom variables */
+
+$omaWrapperPadding: 4rem 0 8rem;
+
+
 // Make svg icons pretty
 .icon {
   /* Lets the icon inherit the text color. */
