+++ omahelsinki/static/css/components/OnboardingSteps/_PersonalInformation.scss
@@ -1,0 +1,3 @@
+.onboarding-personal-information {
+    
+}
