+++ omahelsinki/static/css/components/OnboardingSteps/_Welcome.scss
@@ -1,0 +1,8 @@
+.onboarding-welcome {
+    text-align: center;
+
+    h2 {
+        color: $hel-bus;
+        margin: 2rem 0;
+    }
+}
