+++ omahelsinki/static/css/components/_Main.scss
@@ -27,7 +27,7 @@
 }
 
 .oma-tab-content {
-    padding: 4rem 0 8rem;
+    padding: $omaWrapperPadding;
 }
 
 .oma-tab-container {
