+++ omahelsinki/static/css/components/_Onboarding.scss
@@ -1,0 +1,56 @@
+@import 'OnboardingSteps/Welcome';
+@import 'OnboardingSteps/StepButtons';
+@import 'OnboardingSteps/PersonalInformation';
+
+.oma-onboarding {
+    &-container {
+        display: flex;
+        padding: $omaWrapperPadding;
+        justify-content: center;
+    }
+
+    &-wrapper {
+        height: inherit;
+        width: inherit;
+        background-color: $hel-fog;
+        display: flex;
+    }
+
+    .user-icon {
+        position: absolute;
+        top: -40px;
+        height: 75px;
+        width: 75px;
+        left: calc(50% - 30px);
+        padding: 30px;
+        border-radius: 50%;
+        background-color: $hel-bus;
+        svg {
+            height: 15px;
+            width: 15px;
+            color: $hel-white;
+        }
+    }
+
+    &-steps {
+        padding: 3rem;
+        background-color: $hel-white;
+        h3 {
+            margin-bottom: 2rem;
+        }
+    }
+    
+    .title {
+        display: flex;
+        justify-content: space-between;
+        align-items: center;
+        strong {
+            margin-left: 10px;
+        }
+    }
+
+    display: flex;
+    flex: 1;
+    position: relative;
+    flex-direction: column;
+}
