+++ package.json
@@ -18,6 +18,9 @@
   },
   "homepage": "https://github.com/City-of-Helsinki/omahelsinki#readme",
   "dependencies": {
+    "@fortawesome/fontawesome-svg-core": "^1.2.0",
+    "@fortawesome/free-solid-svg-icons": "^5.1.0",
+    "@fortawesome/react-fontawesome": "^0.1.0",
     "babel-runtime": "^6.23.0",
     "bootstrap": "^4.1.1",
     "jquery": "^1.9.1",
